deployment_services: Wrap in NixOS submodule...
in order to be able to instantiate it multiple times, once per service.
mkDeploymentService was renamed to mkDeploymentUnit in order to better
reflect it's use. It's default values are moved to the NixOS submodule.

This allows to set arbitrary NixOS config values while referencing
each service configuration and will be used for sops.secrets soon